+++ app/jobs/prune_feed_identifications_job.rb
@@ -1,0 +1,7 @@
+class PruneFeedIdentificationsJob < ApplicationJob
+  queue_as :default
+
+  def perform
+    FeedIdentification.obsolete.in_batches(of: 500).delete_all
+  end
+end