+++ db/migrate/20260909143000_add_configuration_digest_to_feed_identifications.rb
@@ -1,0 +1,5 @@
+class AddConfigurationDigestToFeedIdentifications < ActiveRecord::Migration[8.2]
+  def change
+    add_column :feed_identifications, :configuration_digest, :string
+  end
+end